# src/main/resources/static/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: AWS CardDemo API
  description: "REST API for the AWS CardDemo credit-card account management system,\
    \ migrated from COBOL/CICS to Java 25 and Spring Boot. It re-expresses the 17\
    \ legacy online (3270/BMS) screens as JSON request/response endpoints: signon,\
    \ the main and administrative menus, account view/update, card list/view/update,\
    \ transaction list/view/add, transaction reports, bill payment, and administrative\
    \ user management. This is a faithful, behavior-preserving migration and not a\
    \ web UI, terminal emulator, or 3270 renderer; no capabilities beyond the original\
    \ COBOL application are added."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Local development server
- url: /
  description: Relative to the serving host
security:
- basicAuth: []
...
paths:
...
  /api/v1/admin/menu:
    get:
      tags:
      - Admin Menu
      summary: Display the admin menu
      description: First-entry admin menu listing the administrative options (COBOL
        COADM01C initial SEND MAP). Requires the ADMIN role.
      operationId: menu_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminMenuResponse"
        "400":
          description: "Bad Request: the request was malformed or failed input validation\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Unauthorized: authentication is required and was missing or\
            \ invalid (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: "Forbidden: the authenticated user lacks the ADMIN role required\
            \ for this operation (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: "Not Found: the requested resource or record does not exist\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "405":
          description: "Method Not Allowed: the HTTP method is not supported for this\
            \ resource (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: "Internal Server Error: an unexpected error occurred (RFC 7807\
            \ problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
    post:
      tags:
      - Admin Menu
      summary: Submit an admin-menu selection
      description: Reproduces the COADM01C EVALUATE EIBAID routing of the selected
        administrative option.
      operationId: select_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMenuRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminMenuResponse"
        "400":
          description: "Bad Request: the request was malformed or failed input validation\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "Unauthorized: authentication is required and was missing or\
            \ invalid (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: "Forbidden: the authenticated user lacks the ADMIN role required\
            \ for this operation (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: "Not Found: the requested resource or record does not exist\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "405":
          description: "Method Not Allowed: the HTTP method is not supported for this\
            \ resource (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "415":
          description: "Unsupported Media Type: the request body is not application/json\
            \ (RFC 7807 problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: "Internal Server Error: an unexpected error occurred (RFC 7807\
            \ problem+json)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
components:
  schemas:
...
    AdminMenuRequest:
      type: object
      properties:
        option:
          type: string
          maxLength: 2
          minLength: 0
          pattern: "^.{0,2}$"
        action:
          type: string
          enum:
          - ENTER
          - CLEAR
          - PA1
          - PA2
          - PF1
          - PF2
          - PF3
          - PF4
          - PF5
          - PF6
          - PF7
          - PF8
          - PF9
          - PF10
          - PF11
          - PF12
    AdminMenuResponse:
      type: object
      properties:
        transactionName:
          type: string
        title01:
          type: string
        currentDate:
          type: string
        programName:
          type: string
        title02:
          type: string
        currentTime:
          type: string
        menuOptions:
          type: array
          items:
            type: string
        errorMessage:
          type: string
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem detail returned as application/problem+json for
        every error response emitted by the API.
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Not Found
        status:
          type: integer
          format: int32
          example: 404
        detail:
          type: string
          example: The requested account was not found.
        instance:
          type: string
          format: uri
          example: /api/v1/accounts/view
        correlationId:
          type: string
          description: Request correlation id echoed from the X-Correlation-Id header
            / MDC.
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic